src/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.6.3
    # via discord.py
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via aiohttp
beautifulsoup4==4.9.3
    # via bs4
bs4==0.0.1
    # via
    #   -r requirements.in
    #   pydictionary
certifi==2020.11.8
    # via requests
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via
    #   flask
    #   pydictionary
discord-pretty-help==1.2.0
    # via -r requirements.in
discord.py==1.5.1
    # via -r requirements.in
dnspython==2.0.0
    # via -r requirements.in
flask==1.1.2
    # via -r requirements.in
futures==3.1.1
    # via goslate
goslate==1.5.1
    # via pydictionary
idna==2.10
    # via
    #   requests
    #   yarl
itsdangerous==1.1.0
    # via flask
jinja2==2.11.2
    # via flask
markupsafe==1.1.1
    # via jinja2
multidict==4.7.6
    # via
    #   aiohttp
    #   yarl
pydictionary==2.0.1
    # via -r requirements.in
pymongo==3.11.1
    # via -r requirements.in
python-dotenv==0.15.0
    # via -r requirements.in
pytz==2020.4
    # via -r requirements.in
requests==2.25.0
    # via
    #   -r requirements.in
    #   pydictionary
soupsieve==2.0.1
    # via beautifulsoup4
urllib3==1.26.2
    # via requests
werkzeug==1.0.1
    # via flask
yarl==1.5.1
    # via aiohttp
